+++ activemodel/CHANGELOG.md
@@ -1,3 +1,38 @@
+*   Introduce `ActiveModel::Type::Model` for model attributes
+
+    ```ruby
+    class Article
+      include ActiveModel::Attributes
+
+      attribute :author, class_name: "Person"
+      attribute :comments, class_name: "Comment", array: true
+    end
+
+    class Person
+      include ActiveModel::API
+      include ActiveModel::Attributes
+
+      attribute :name, :string
+    end
+
+    class Comment
+      include ActiveModel::API
+      include ActiveModel::Attributes
+
+      attribute :body, :string
+    end
+
+    article = Article.new
+
+    article.author = { name: "Arthur" }
+    article.author.name # => "Arthur"
+
+    article.comments = [{ body: "Hello" }, { body: "Goodbye" }]
+    article.comments.map(&:body) # => ["Hello", "Goodbye"]
+    ```
+
+    *Sean Doyle*
+
 *   Fix a bug where type casting of string to `Time` and `DateTime` doesn't
     calculate minus minute value in TZ offset correctly.